# openapi.json
{
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Machine-readable error code",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Optional additional details",
                "type": "object"
              },
              "message": {
                "description": "Human-readable error message",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error",
          "trace_id"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Reeverb API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/projects/{id}/testimonials": {
      "get": {
        "operationId": "list_testimonials",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "TagResponse": {
                      "properties": {
                        "color": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "project_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "project_id",
                        "name"
                      ],
                      "type": "object"
                    },
                    "TestimonialResponse": {
                      "properties": {
                        "author_avatar_url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "author_company": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "author_email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "author_name": {
                          "type": "string"
                        },
                        "author_title": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "author_url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "content": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "is_approved": {
                          "type": "boolean"
                        },
                        "is_featured": {
                          "type": "boolean"
                        },
                        "language": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "project_id": {
                          "type": "string"
                        },
                        "rating": {
                          "format": "int16",
                          "maximum": 32767,
                          "minimum": -32768,
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "sentiment": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "sentiment_score": {
                          "format": "float",
                          "type": [
                            "number",
                            "null"
                          ]
                        },
                        "source": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "source_id": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "source_platform": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "source_url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "tags": {
                          "items": {
                            "$ref": "#/$defs/TagResponse"
                          },
                          "type": "array"
                        },
                        "transcription": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "type": {
                          "type": "string"
                        },
                        "updated_at": {
                          "type": "string"
                        },
                        "video_duration_seconds": {
                          "format": "int32",
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "video_thumbnail_url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "video_url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "project_id",
                        "type",
                        "author_name",
                        "is_approved",
                        "is_featured",
                        "tags",
                        "created_at",
                        "updated_at"
                      ],
                      "type": "object"
                    }
                  },
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "$ref": "#/$defs/TestimonialResponse"
                  },
                  "title": "Array_of_TestimonialResponse",
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "User does not own this project"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Testimonial not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        },
        "summary": "List testimonials"
      },
      "post": {
        "operationId": "create_testimonial",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "User does not own this project"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Testimonial not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        },
        "summary": "Create testimonial"
      }
    },
...
    "/api/v1/testimonials/{id}/approve": {
      "post": {
        "operationId": "approve_testimonial",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "TagResponse": {
                      "properties": {
                        "color": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "project_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "project_id",
                        "name"
                      ],
                      "type": "object"
                    }
                  },
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "properties": {
                    "author_avatar_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "author_company": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "author_email": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "author_name": {
                      "type": "string"
                    },
                    "author_title": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "author_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "content": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "is_approved": {
                      "type": "boolean"
                    },
                    "is_featured": {
                      "type": "boolean"
                    },
                    "language": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "project_id": {
                      "type": "string"
                    },
                    "rating": {
                      "format": "int16",
                      "maximum": 32767,
                      "minimum": -32768,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "sentiment": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "sentiment_score": {
                      "format": "float",
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "source": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "source_id": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "source_platform": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "source_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "tags": {
                      "items": {
                        "$ref": "#/$defs/TagResponse"
                      },
                      "type": "array"
                    },
                    "transcription": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "type": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "video_duration_seconds": {
                      "format": "int32",
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "video_thumbnail_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "video_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "project_id",
                    "type",
                    "author_name",
                    "is_approved",
                    "is_featured",
                    "tags",
                    "created_at",
                    "updated_at"
                  ],
                  "title": "TestimonialResponse",
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "User does not own this project"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Testimonial not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        },
        "summary": "Approve testimonial"
      }
    },
...
  }
}